Reorganize bug tests into 5 focused test files grouped by server functionality and expected fixes.
Learned: Merging tests by server code path and creating separate files by functional area makes test suite more maintainable than one large file.
